+++ R/write.R
@@ -75,9 +75,10 @@ write_delim <- function(x, file, delim = " ", na = "NA", append = FALSE,
   }
 
   stopifnot(is.data.frame(x))
+  check_list_columns(x)
 
   x_out <- x
-  x[] <- lapply(names(x), function(i) output_column(x[[i]], i))
+  x[] <- lapply(x, output_column)
   stream_delim(x, file, delim = delim, col_names = col_names, append = append,
     na = na, quote_escape = quote_escape, eol = eol)
 
@@ -124,6 +125,7 @@ write_excel_csv <- function(x, file, na = "NA", append = FALSE,
   }
 
   stopifnot(is.data.frame(x))
+  check_list_columns(x)
 
   x_out <- x
   datetime_cols <- vapply(x, inherits, logical(1), "POSIXt")
@@ -146,6 +148,9 @@ write_excel_csv2 <- function(x, file, na = "NA", append = FALSE,
     deprecate_warn("1.4.0", "write_excel_csv2(path = )", "write_excel_csv2(file = )")
     file <- path
   }
+
+  stopifnot(is.data.frame(x))
+  check_list_columns(x)
 
   x_out <- x
   x <- change_decimal_separator(x, decimal_mark = ",")
@@ -203,6 +208,7 @@ write_tsv <- function(x, file, na = "NA", append = FALSE, col_names = !append,
 format_delim <- function(x, delim, na = "NA", append = FALSE,
                          col_names = !append, quote_escape = "double", eol = "\n") {
   stopifnot(is.data.frame(x))
+  check_list_columns(x)
 
   x[] <- lapply(x, output_column)
   res <- stream_delim(df = x, file = NULL, delim = delim, col_names = col_names, append = append, na = na, quote_escape = quote_escape, eol = eol)
@@ -264,11 +270,6 @@ output_column.POSIXt <- function(x, name) {
   format(x, "%Y-%m-%dT%H:%M:%OSZ", tz = "UTC", justify = "none")
 }
 
-#' @export
-output_column.list <- function(x, name) {
-  stop("Flat files can't store the list column `", name, "`", call. = FALSE)
-}
-
 stream_delim <- function(df, file, append = FALSE, bom = FALSE, ..., quote_escape, eol) {
   quote_escape <- standardise_escape(quote_escape)
 
@@ -310,4 +311,14 @@ standardise_escape <- function(x) {
   escape <- match.arg(tolower(x), names(escape_types))
 
   escape_types[escape]
+}
+
+check_list_columns <- function(x) {
+  is_list_column <- vapply(x, inherits, logical(1), "list")
+  if (any(is_list_column)) {
+    cli_block(type = rlang::abort, {
+      cli::cli_text("`x` must not contain list columns:")
+      cli::cli_alert_danger("list columns at index(s): {paste0(which(is_list_column), collapse = '\n')}")
+    })
+  }
 }